Moved SpellEffect to Abstracts folder
