+++ src/main/java/com/myuuiii/empirewandplus/Data/SpellEffectHashMap.java
@@ -7,7 +7,7 @@ import com.myuuiii.empirewandplus.SpellEffects.Lightning.LightningEffect;
 import com.myuuiii.empirewandplus.SpellEffects.Pulse.EmpirePulseEffect;
 import com.myuuiii.empirewandplus.SpellEffects.Pulse.FirePulseEffect;
 import com.myuuiii.empirewandplus.SpellEffects.Smite.SmiteEffect;
-import com.myuuiii.empirewandplus.SpellEffects.SpellEffect;
+import com.myuuiii.empirewandplus.Abstracts.SpellEffect;
 import com.myuuiii.empirewandplus.SpellEffects.Wave.BloodWaveEffect;
 import com.myuuiii.empirewandplus.SpellEffects.Wave.FlameWaveEffect;
 import com.myuuiii.empirewandplus.SpellEffects.Wave.PoisonWaveEffect;
