+++ src/main/java/com/myuuiii/empirewandplus/SpellEffects/Pulse/FirePulseEffect.java
@@ -2,7 +2,7 @@ package com.myuuiii.empirewandplus.SpellEffects.Pulse;
 
 import com.myuuiii.empirewandplus.Abstracts.Spell;
 import com.myuuiii.empirewandplus.EmpireWandPlus;
-import com.myuuiii.empirewandplus.SpellEffects.SpellEffect;
+import com.myuuiii.empirewandplus.Abstracts.SpellEffect;
 import com.myuuiii.empirewandplus.Data.SpellEntityLists;
 import com.myuuiii.empirewandplus.Data.SpellNames;
 import org.bukkit.Color;
